// src/types/product-calculation-report/productCalculationReportList.type.ts
export interface IproductCalculationReportList {
  product_name: string;
  sku: string;
  boston_beer: string;
  bacardi_grey_goose: string;
  cyxtera: string;
  simpliSafe: string;
  the_home_depot: string;
  unit1: string;
  USAA_alpha_healthy_points: string;
  Usaa_Claims_Apparel: string;
  Usaa_Healthy_Points: string;
  Usaa_Punchout: string;
  Porsche: string;
  Skadden_Holiday_Store: string;
  Stryker: string;
  USAA_NEO_Shirts: string;
}